backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
aiosqlite==0.20.0
sqlalchemy==2.0.36
alembic==1.13.3
pytest==8.3.0
ruff==0.6.0
black==24.8.0
mypy==1.12.0
jose==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-docx==1.1.0

